Typo in return type
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/rgl/pkg/rgl@730 79a1b648-8b68-254f-b20b-c0b50f5137bf

+++ src/api.cpp
@@ -1212,7 +1212,7 @@ double getCex()
   Device* device;
   
   if (deviceManager && (device = deviceManager->getCurrentDevice())) {
-    int result = device->getRGLView()->getFontCex();
+    double result = device->getRGLView()->getFontCex();
     CHECKGLERROR;  
     return result;
   } else
